pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trader"
version = "0.1.0"
description = "Automated swing trading system — mean reversion dip-buy strategy"
requires-python = ">=3.11,<3.13"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "pyarrow>=15.0",
    "yfinance>=0.2.50",
    "matplotlib>=3.8",
    "python-dotenv>=1.0",
    "vectorbt>=0.26",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
